move as_lurk_boolean to the right place
